+++ orc8r/cloud/configs/bootstrapper.yml
@@ -9,3 +9,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+tokenTimeoutDurationInMinutes: 30
